Improve app data loading and prepare for Android 15

+++ app-common-io/src/main/java/eu/darken/sdmse/common/pkgs/pkgops/ipc/PkgOpsClient.kt
@@ -94,14 +94,14 @@ class PkgOpsClient @AssistedInject constructor(
      * Can fail if the amount of packages exceeds the IPC buffer size.
      * android.os.DeadObjectException: Transaction failed on small parcel; remote process probably died
      */
-    fun getInstalledPackagesAsUser(flags: Int, userHandle: UserHandle2): List<PackageInfo> = try {
+    fun getInstalledPackagesAsUser(flags: Long, userHandle: UserHandle2): List<PackageInfo> = try {
         connection.getInstalledPackagesAsUser(flags, userHandle.handleId)
     } catch (e: Exception) {
         log(TAG, ERROR) { "getInstalledPackagesAsUser(flags=$flags, userHandle=$userHandle) failed: ${e.asLog()}" }
         throw fakeIOException(e.getRootCause())
     }
 
-    fun getInstalledPackagesAsUserStream(flags: Int, userHandle: UserHandle2): List<PackageInfo> = try {
+    fun getInstalledPackagesAsUserStream(flags: Long, userHandle: UserHandle2): List<PackageInfo> = try {
         connection.getInstalledPackagesAsUserStream(flags, userHandle.handleId).toPackageInfos()
     } catch (e: Exception) {
         log(
